Term::url() - ensure id is passed as int to get_term_link
This preserves the current behavior of returning an empty term in the event that the model’s id is null.

This is because `get_term_link()` uses `get_term()` for integer term references, otherwise it uses `get_term_by()` for a slug match. Recent changes in trunk in `get_term_by()` have caused this to break when passing null, and returning the Uncategorized term rather than a WP_Error.